requirements.txt
pygame
opencv-python
google-generativeai
pyaudio
SpeechRecognition
numpy
pillow
pyautogui
requests
